requirements.txt
python-telegram-bot==21.3
aiohttp==3.10.5
apscheduler==3.10.4
